requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
idna==3.3
requests==2.27.1
urllib3==1.26.9
